+++ tests/c/c_node_link_events/src/main.c
@@ -108,7 +108,7 @@ static int register_yuno_and_more(void)
      *------------------------------*/
     set_expected_results(
         APP_NAME,
-        json_pack("[{s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}]",
+        json_pack("[{s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}, {s:s}]",
             "msg", "Starting yuno",
             "msg", "Creating __timeranger2__.json",
             "msg", "Creating TreeDB schema file",
@@ -120,6 +120,9 @@ static int register_yuno_and_more(void)
             "msg", "Playing yuno",
             "msg", "fkey reference: parent node not found",
             "msg", "fkey reference: its hook does not link into this column",
+            "msg", "fkey reference: parent node not found",
+            "msg", "fkey reference: parent node not found",
+            "msg", "Node already exists",
             "msg", "with_link_events changed",
             "msg", "with_link_events changed",
             "msg", "Topic name not found in treedbs",